Наведение красоты кода, оптимизация, поставщики, фотки и еще много чего

+++ SparesBase/Dobavlenie/Forms/EmployeesForm.cs
@@ -5,6 +5,8 @@ namespace SparesBase.Forms
 {
     public partial class EmployeesForm : Form
     {
+        #region Конструкторы
+        
         // Конструктор
         public EmployeesForm()
         {
@@ -12,33 +14,59 @@ namespace SparesBase.Forms
             FillEmployees();
         }
 
+        #endregion Конструкторы
 
+
+
+        #region Методы
+        
         // Заполнение сотрудников
         private void FillEmployees()
         {
             dgv.Rows.Clear();
-            DataTable employees = DatabaseWorker.SqlSelectQuery("SELECT a.id, a.FirstName, a.LastName, a.SecondName, a.Login, c.City, a.Phone, a.Email, a.Admin, o.id, o.Name, o.Site, o.Telephone, oc.City  FROM Accounts a LEFT JOIN Cities c ON c.id=a.CityId LEFT JOIN Organizations o ON o.id=a.OrganizationId LEFT JOIN Cities oc ON oc.id=o.CityId WHERE(OrganizationId=" + EnteredUser.OrganizationId + ")");
+            DataTable employees = DatabaseWorker.SqlSelectQuery("SELECT " +
+                "a.id, " +
+                "a.FirstName, " +
+                "a.LastName, " +
+                "a.SecondName, " +
+                "a.Login, " +
+                "a.Password, " +
+                "o.id, " +
+                "o.Name, " +
+                "o.Site, " +
+                "o.Telephone, " +
+                "oc.City, " +
+                "ac.City, " +
+                "a.Phone, " +
+                "a.Email, " +
+                "a.Admin " +
+                "FROM Accounts a " +
+                "LEFT JOIN Organizations o ON o.id = a.OrganizationId " +
+                "LEFT JOIN Cities ac ON ac.id = a.CityId " +
+                "LEFT JOIN Cities oc ON oc.id = o.CityId " +
+                "WHERE(OrganizationId = " + EnteredUser.Organization.Id + ")");
+
             foreach (DataRow row in employees.Rows)
             {
                 Account account = new Account(
-                     int.Parse(row.ItemArray[0].ToString()),
-                     row.ItemArray[1].ToString(),
-                     row.ItemArray[2].ToString(),
-                     row.ItemArray[3].ToString(),
-                     row.ItemArray[4].ToString(),
-                     row.ItemArray[5].ToString(),
-                     row.ItemArray[6].ToString(),
-                     row.ItemArray[7].ToString(),
-                     row.ItemArray[8].ToString() == "1" ? true : false,
-                     null);
-
-                Organization org = new Organization(
-                    int.Parse(row.ItemArray[9].ToString()),
-                    row.ItemArray[10].ToString(),
+                    int.Parse(row.ItemArray[0].ToString()),
+                    row.ItemArray[1].ToString(),
+                    row.ItemArray[2].ToString(),
+                    row.ItemArray[3].ToString(),
+                    row.ItemArray[4].ToString(),
+                    null,
                     row.ItemArray[11].ToString(),
                     row.ItemArray[12].ToString(),
-                    row.ItemArray[13].ToString(),                                    
-                    account);
+                    row.ItemArray[13].ToString(),
+                    row.ItemArray[14].ToString() == "1" ? true : false);
+
+                Organization org = new Organization(
+                        int.Parse(row.ItemArray[6].ToString()),
+                        row.ItemArray[7].ToString(),
+                        row.ItemArray[8].ToString(),
+                        row.ItemArray[9].ToString(),
+                        row.ItemArray[10].ToString(),
+                        null);
                 account.Organization = org;
 
                 dgv.Rows.Add(account.LastName, account.FirstName, account.SecondName);
@@ -46,7 +74,12 @@ namespace SparesBase.Forms
             }
         }
 
+        #endregion Методы
 
+
+
+        #region События
+        
         // Двойной клик на сотруднике
         private void dgv_CellDoubleClick(object sender, DataGridViewCellEventArgs e)
         {
@@ -55,7 +88,7 @@ namespace SparesBase.Forms
         }
 
         // Добавление сотрудника
-        private void tsmiAdd_Click(object sender, System.EventArgs e)
+        private void Add_Click(object sender, System.EventArgs e)
         {
             RegistrationForm reg = new RegistrationForm();
             reg.ShowDialog();
@@ -63,7 +96,7 @@ namespace SparesBase.Forms
         }
 
         // Редактирование сотрудника
-        private void tsmiEdit_Click(object sender, System.EventArgs e)
+        private void Edit_Click(object sender, System.EventArgs e)
         {
             AccountEditor ae = new AccountEditor((Account)dgv.CurrentRow.Tag);
             ae.ShowDialog();
@@ -71,19 +104,30 @@ namespace SparesBase.Forms
         }
 
         // Удаление сотрудника
-        private void tsmiDelete_Click(object sender, System.EventArgs e)
+        private void Delete_Click(object sender, System.EventArgs e)
         {
             if (dgv.SelectedRows.Count > 0)
             {
                 Account account = (Account)dgv.CurrentRow.Tag;
                 if (!account.Admin)
                 {
-                    DatabaseWorker.SqlQuery("DELETE FROM Accounts WHERE(id=" + account.Id + ")");
-                    FillEmployees();
+                    if (MessageBox.Show("Вы действительно хотите удалить сотрудника " + 
+                        account.LastName + " " +
+                        account.FirstName + " " +
+                        account.SecondName + "?", 
+                        "Удаление сотрудника", 
+                        MessageBoxButtons.YesNo, 
+                        MessageBoxIcon.Question) == DialogResult.Yes)
+                    {
+                        DatabaseWorker.SqlQuery("DELETE FROM Accounts WHERE(id = " + account.Id + ")");
+                        FillEmployees();
+                    }
                 }
                 else
                     MessageBox.Show("Удаление админа произвести невозможно");   
             }
         }
+
+        #endregion События
     }
 }
